Lab 6/requirements.txt
Adafruit-Blinka==6.6.0
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-rgb-display==3.10.6
Adafruit-PlatformDetect==3.8.0
Adafruit-PureIO==1.1.8
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-speech==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
libcst==0.3.18
mypy-extensions==0.4.3
packaging==20.9
paho-mqtt==1.5.1
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyftdi==0.52.9
pyparsing==2.4.7
pyserial==3.5
pyttsx3==2.90
pytz==2021.1
pyusb==1.1.1
PyYAML==5.4.1
requests==2.25.1
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
rsa==4.7.2
six==1.15.0
SpeechRecognition==3.8.1
sysv-ipc==1.1.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
